requirements.txt
ag2==0.9.1.post0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.1.post0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
camel-ai==0.2.58
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
croniter==6.0.0
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring-parser==0.15
executing==2.2.0
Flask==3.1.1
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.21.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.14
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.8.1
numpy==2.2.6
oauthlib==3.2.2
openai==1.78.1
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.27.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.52.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
tenacity==8.5.0
termcolor==3.1.0
tiktoken==0.7.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
openpyxl==3.1.5
